refactor to shared package
