+++ packages/engine/src/index.ts
@@ -1,5 +1,3 @@
 export * from "./engine";
 export * from "./inputManager";
 export * from "./api";
-export * from "./utils";
-export * from "./scoring";
