+++ packages/frontend/src/plugins/injectGlobalCssVars.ts
@@ -1,4 +1,4 @@
-import { NoteSkin } from "@packages/types";
+import { NoteSkin } from "@packages/shared";
 
 export function injectStylesheet(style: string, id: string) {
   let el = document.querySelector<HTMLStyleElement>(`#${id}`);
