+++ packages/frontend/src/types.ts
@@ -2,7 +2,7 @@ import {
   ParsedHoldNoteChart,
   ParsedTapNoteChart,
 } from "@packages/chart-parser";
-import type { BaseSong, Difficulty } from "@packages/types";
+import type { BaseSong, Difficulty } from "@packages/shared";
 
 export interface Song extends BaseSong {
   order: number;
