+++ packages/game-data/package.json
@@ -19,7 +19,7 @@
   "devDependencies": {
     "@packages/chart-parser": "0.0.0-development",
     "@packages/engine": "0.0.0-development",
-    "@packages/types": "0.0.0-development",
+    "@packages/shared": "0.0.0-development",
     "@types/cookie-parser": "^1.4.3",
     "@types/cors": "^2.8.12",
     "@types/express": "^4.17.13",
