+++ packages/shared/src/scoring.spec.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect } from "vitest";
-import type { World, EngineNote, TimingWindow } from "./engine";
+import type { World, EngineNote, TimingWindow } from "@packages/engine";
 import { createWorld } from "../test/utils";
 import { Summary, summarizeResults, SummaryData } from "./scoring";
 import { extractNotesFromWorld } from "./utils";
